changed m0 model exception handling manualy
Instead of seting a pending interupt go derectly into exceptional state.

for our work we need to ensure that any non excepitonal state transition
incresses the clock count monotonically.

The problem with original model is that undefined instructions and SVC
will increes the clock count by 0, or ARB respectivaly and do not lead
to an exceptional state.

This should not break the decompiler since the decompiler does not
handle undefined instructions and interupts tobegin with.

Holmake builds, and m0_decomp_demo.sml works.